requirements.txt
pandas~=2.2.3
streamlit~=1.48.0
numpy~=2.1.3
streamlit-extras==0.5.0
gspread == 6.2.1
google-genai==1.33